template_service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7